[refactor] refactor the react component
