+++ components/Common/Styled/animations.ts
@@ -1,0 +1,21 @@
+import { keyframes } from 'styled-components';
+
+// Common animations
+export const fadeIn = keyframes`
+  from { opacity: 0; transform: translateY(10px); }
+  to { opacity: 1; transform: translateY(0); }
+`;
+
+export const bounce = keyframes`
+  0%, 100% { transform: translateY(0); }
+  50% { transform: translateY(-10px); }
+`;
+
+export const spin = keyframes`
+  to { transform: rotate(360deg); }
+`;
+
+export const slideUp = keyframes`
+  from { transform: translateY(20px); opacity: 0; }
+  to { transform: translateY(0); opacity: 1; }
+`;
