+++ components/Common/Styled/buttons.ts
@@ -1,0 +1,64 @@
+import styled from 'styled-components';
+import { Loader2 } from 'lucide-react';
+import { spin } from './animations';
+
+export const IconButton = styled.button`
+  width: 40px;
+  height: 40px;
+  border-radius: 8px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  background-color: transparent;
+  color: ${({ theme }) => theme.colors.textSecondary};
+  transition: all ${({ theme }) => theme.transitions.base};
+  border: none;
+  cursor: pointer;
+
+  &:hover {
+    background-color: ${({ theme }) => theme.colors.surfaceHover};
+    color: ${({ theme }) => theme.colors.primary};
+  }
+`;
+
+export const PrimaryButton = styled.button`
+  width: 100%;
+  padding: ${({ theme }) => theme.spacing.md};
+  background: linear-gradient(135deg, #2563eb 0%, #1d4ed8 100%);
+  color: white;
+  border: none;
+  border-radius: ${({ theme }) => theme.borderRadius.lg};
+  font-size: ${({ theme }) => theme.fontSizes.base};
+  font-weight: 600;
+  transition: all ${({ theme }) => theme.transitions.base};
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: ${({ theme }) => theme.spacing.sm};
+  box-shadow: ${({ theme }) => theme.shadows.lg};
+  cursor: pointer;
+
+  &:hover:not(:disabled) {
+    transform: translateY(-2px);
+    box-shadow: 0 10px 25px rgba(37, 99, 235, 0.3);
+  }
+
+  &:disabled {
+    opacity: 0.6;
+    cursor: not-allowed;
+    background: #94a3b8;
+  }
+`;
+
+export const Spinner = styled.div`
+  width: 20px;
+  height: 20px;
+  border: 3px solid rgba(255, 255, 255, 0.3);
+  border-radius: 50%;
+  border-top-color: white;
+  animation: ${spin} 1s ease-in-out infinite;
+`;
+
+export const LoadingSpinner = styled(Loader2)`
+  animation: ${spin} 1s linear infinite;
+`;
