+++ components/WritingStudio/Styled/index.ts
@@ -1,0 +1,6 @@
+// Export all WritingStudio-specific styled components
+export * from './EditorPanel';
+export * from './OutputPanel';
+export * from './IELTSScore';
+export * from './Vocabulary';
+export * from './WordPopover';
